+++ example/lib/main_container.dart
@@ -1,0 +1,5 @@
+import 'package:view_model_kit/view_model_kit.dart';
+
+class MainContainer extends BaseContainer {
+  late final mainCounter = create(0);
+}