+++ lib/src/rx_container.dart
@@ -1,0 +1,9 @@
+part of view_model_kit;
+
+abstract class BaseContainer {
+  BaseContainer() {
+    onCreate();
+  }
+
+  void onCreate() {}
+}