+++ CHANGELOG.md
@@ -1,8 +1,11 @@
+## 0.0.4
+- add BaseContainer(can use R<V> or RList<V> without ViewModel)
+
 ## 0.0.3
-- change RList<V>.value to public.
+- change [RList<V>.value] to public.
 
 ## 0.0.2
-- add R<V>.observe and RList<V>.observe method.
+- add [R<V>.observe] and [RList<V>.observe] method.
 - add SelectBuilder Widget.
 
 ## 0.0.1